Remove repository backward compatibility code
The code added to support backward compatibility with the old version
of the Nimble cache packages repository is removed because migration
functionality to the new cache format will not be implemented. The
reason for this is that checksums of the packages must be calculated on
package download including all files, instead of only installed ones.
Otherwise, some files which are not installed could be used in the build
of a binary package executable and if so the checksum will not describe
the package correctly since the executable is also installed.

Additionally, this commit fixes some tests.

Related to nim-lang/nimble#127